Several bugs corrected. Now UART output tic works (but by changing pin assigned RP8 to RP7). configUart simplified.

+++ encoder-angle-mapping.h
@@ -5,4 +5,4 @@
 /** Doc: see http://code.google.com/p/jig-torque-positioning   **/
 /****************************************************************/
 
-float getBestAngle(long encPos);
+int getBestAngle(long encPos);
